se corrigieron los métodos de Usuario.Service y UsuarioController.java y se modificó el tipo de dato bike_id a int

+++ parking_ud/src/main/java/org/example/parking_ud/dao/Bicycle.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 public class Bicycle {
     @Id
     @Column(name = "bike_id", nullable = false)
-    public Short id;
+    public int id;
 
     @Size(max = 50)
     @NotNull
